requirements.txt
python-telegram-bot==22.0
python-dotenv==1.1.0
faster-whisper==1.1.1
claude-code-sdk>=0.1.0
